chore(SystemPolicyCard): Refactor to function component
