+++ src/PresentationalComponents/SystemPolicyCard/SystemPolicyCard.test.js
@@ -1,8 +1,6 @@
-import React from 'react';
-import { render } from 'enzyme';
-import toJson from 'enzyme-to-json';
+import { render } from '@testing-library/react';
+import { queryByText } from '@testing-library/dom';
 import SystemPolicyCard from './SystemPolicyCard';
-import { IntlProvider } from 'react-intl';
 
 describe('SystemPolicyCard component', () => {
   const currentTime = new Date();
@@ -23,13 +21,20 @@ describe('SystemPolicyCard component', () => {
     },
   };
 
-  it('should render policy', () => {
-    const wrapper = render(
-      <IntlProvider locale={navigator.language}>
-        <SystemPolicyCard policy={policy} />
-      </IntlProvider>
-    );
-    expect(toJson(wrapper)).toMatchSnapshot();
+  it('should render non compliant policy', () => {
+    const { container } = render(<SystemPolicyCard policy={policy} />);
+
+    expect(queryByText(container, 'Not compliant')).not.toBeNull();
+  });
+
+  it('should render compliant policy', () => {
+    const compliantPolicy = {
+      ...policy,
+      compliant: true,
+    };
+    const { container } = render(<SystemPolicyCard policy={compliantPolicy} />);
+
+    expect(queryByText(container, 'Compliant')).not.toBeNull();
   });
 
   it('should render an unsupported policy', () => {
@@ -37,11 +42,12 @@ describe('SystemPolicyCard component', () => {
       ...policy,
       supported: false,
     };
-    const wrapper = render(
-      <IntlProvider locale={navigator.language}>
-        <SystemPolicyCard policy={unsupportedPolicy} />
-      </IntlProvider>
+    const { container } = render(
+      <SystemPolicyCard policy={unsupportedPolicy} />
     );
-    expect(toJson(wrapper)).toMatchSnapshot();
+
+    expect(
+      queryByText(container, 'Unsupported SSG version (0.1.45)')
+    ).not.toBeNull();
   });
 });
